Update: SignalGeneratorTest Set the magnitude to float.

+++ tests/core/signal_generator_test.cpp
@@ -127,14 +127,16 @@ TEST_F(SignalGeneratorTest, OverlaySignal) {
   EXPECT_EQ(m_audioFile->getLengthInSeconds(), TEST_AUDIO_LENGTH);
 
   // Generate signal.
-  m_signalGenerator->generateSignal(32, 1 / 3);
+  m_signalGenerator->generateSignal(32, (float)1 / (float)3);
 
   // Overlay 2 layers of new signal.
-  m_signalGenerator->overlaySignal(64, 1 / 3);
-  m_signalGenerator->overlaySignal(128, 1 / 3);
+  m_signalGenerator->overlaySignal(64, (float)1 / (float)3);
+  m_signalGenerator->overlaySignal(128, (float)1 / (float)3);
 
   // Test signal.
-  std::vector<FreqComponent> freq{{32, 1 / 3}, {64, 1 / 3}, {128, 1 / 3}};
+  std::vector<FreqComponent> freq{{32, (float)1 / (float)3},
+                                  {64, (float)1 / (float)3},
+                                  {128, (float)1 / (float)3}};
   EXPECT_TRUE(isFrequency(freq));
 }
